Moving __tests__ folder out of src and modifying prop name
