+++ package.json
@@ -16,7 +16,7 @@
     "lint": "npm run lint:js && npm run lint:prettier",
     "clean": "rm -r dist",
     "precommit": "lint-staged",
-    "build": "cross-env NODE_ENV=production babel src --out-dir dist --ignore \"**/*.test.js\"",
+    "build": "cross-env NODE_ENV=production babel src --out-dir dist --ignore \"**/*.test.js\" --copy-files",
     "prepublishOnly": "npm run lint && npm run test && npm run clean && npm run build"
   },
   "lint-staged": {
